% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{add_pseudo_barcode}
\alias{add_pseudo_barcode}
\title{add_pseudo_barcode}
\usage{
add_pseudo_barcode(barcode)
}
\arguments{
\item{barcode}{a barcode obtained from the compute_barcode or compute_barcode_multi functions}
}
\value{
returns a matrix with a barcode for which a pseudobarcoded has been added to each demiension.
}
\description{
This function adds a pseudobarcode to each dimension of the generated barcode.
}
\examples{
\dontrun{
add_pseudo_barcode(barcode)
}
}
