% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{assign_rounds_for_bc}
\alias{assign_rounds_for_bc}
\title{assign_rounds_for_bc}
\usage{
assign_rounds_for_bc(path_to_joined_five)
}
\arguments{
\item{path_to_joined_five}{Paht to the joined_five file.}
}
\value{

}
\description{
Generate randmly sampled round to carry out the project in different steps.
}
\examples{
\dontrun{
assign_rounds_for_bc(path_to_joined_five)
}

}
