% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulate_PDB.R
\name{aux_for_extract_multiple_alpha}
\alias{aux_for_extract_multiple_alpha}
\title{aux_for_extract_multiple_alpha}
\usage{
aux_for_extract_multiple_alpha(struc)
}
\arguments{
\item{struc}{A particular pdb structure loaded using the read.pdb function from the bio3d package.}
}
\value{
Returns a matrix containing the coordinates of the alpha carbons.
}
\description{
Auxiliary function for the extract_multiple_alpha function. Given a pdb structure it extracts the coordinats from its alpha carbons.
}
\examples{
\dontrun{
aux_for_extract_multiple_alpha(struc)
}
}
