% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{aux_fun_complete_matrices}
\alias{aux_fun_complete_matrices}
\title{aux_fun_complete_matrices}
\usage{
aux_fun_complete_matrices(x)
}
\arguments{
\item{x}{upper diagonal matrix to complete.}
}
\value{

}
\description{
aux_fun_complete_matrices
}
\examples{
\dontrun{
aux_fun_complete_matrices(x)
}
}
