% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{aux_function}
\alias{aux_function}
\title{aux_func}
\usage{
aux_function(x)
}
\arguments{
\item{x}{a df with the same column structure than joined_5.}
}
\value{

}
\description{
auxiliar function to carry out the random sampling to generate different project rounds.
}
\examples{
\dontrun{
aux_function(x)
}

}
