% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{aux_function_samp_b}
\alias{aux_function_samp_b}
\title{aux_function_samp_b}
\usage{
aux_function_samp_b(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Auxiliar function for sample_structures_b
}
\examples{
\dontrun{
aux_function_samp_b(x)
}
}
