% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsing_Vina_Logs.R
\name{aux_same_drug_diff_struc}
\alias{aux_same_drug_diff_struc}
\title{aux_same_drug_diff_struc}
\usage{
aux_same_drug_diff_struc(x)
}
\arguments{
\item{x}{dataframe of different structures docked to the same ligand.}
}
\value{

}
\description{
Auxiliary function to compute diferences in docking energies between different structures docked to the same drug.
}
\examples{
\dontrun{
aux_same_drug_diff_struc(x)
}
}
