% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{complete_matrices}
\alias{complete_matrices}
\title{complete_matrices}
\usage{
complete_matrices(sim_mat_list)
}
\arguments{
\item{sim_mat_list}{list of similarity matrices derived from compute_sim_one_dir_bar or compute_sim_two_dir_bar}
}
\value{

}
\description{
complete_matrices
}
\examples{
\dontrun{
complete_matrices(sim_mat_list)
}
}
