% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_bc_and_save}
\alias{compute_bc_and_save}
\title{compute_bc_and_save}
\usage{
compute_bc_and_save(dir_in, dir_out, n_alpha_filt = 1000)
}
\arguments{
\item{dir_in}{Directory where the biological assembly pdb structures were stored.}

\item{dir_out}{Output directory}

\item{n_alpha_filt}{Maximum number of alpha carbons allowed.}
}
\value{

}
\description{
Given a directory where pdb structures are stored and an output directory the function reads the pdb structures if the number of alopha carbons is smaller than a specificed trheshold it computes the barcodes for the struecture and save them in the output directory.
}
\examples{
\dontrun{
compute_bc_and_save(dir_in,ir_bc,alpha_filt = 1000)
}
}
