% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_bc_round}
\alias{compute_bc_round}
\title{compute_bc_round}
\usage{
compute_bc_round(x, round_n, dir_bc)
}
\arguments{
\item{x}{a dataframe with the whole database information.}

\item{round_n}{the number of the round to be computed.}

\item{dir_bc}{the directory where barcode files must be stored.}
}
\value{

}
\description{
compute_bc_round
}
\examples{
\dontrun{
compute_bc_round(x,round_n,dir_bc)
}
}
