% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_bc_round_multi}
\alias{compute_bc_round_multi}
\title{compute_bc_round_multi}
\usage{
compute_bc_round_multi(x, dir_bc)
}
\arguments{
\item{x}{a dataframe with the whole database information.}

\item{dir_bc}{the directory where barcode files must be stored.}

\item{round_n}{the number of the round to be computed.}
}
\value{

}
\description{
A multiple thread version of the previous function.
}
\examples{
\dontrun{
compute_bc_round(x,dir_bc)
}
}
