% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docking_Functions.R
\name{compute_by_square}
\alias{compute_by_square}
\title{compute_by_square}
\usage{
compute_by_square(
  bio_assay_pdbqt,
  drug_pdbqt_paths,
  output_path,
  bio_assay_paths,
  exhaust = 40,
  save_control_matrix_path,
  verbose = TRUE
)
}
\arguments{
\item{bio_assay_pdbqt}{vector with the absolute paths to the pdbqt files of the proteic structures we want to dock.}

\item{drug_pdbqt_paths}{vector with the absolute paths to the drug pdbqt files.}

\item{output_path}{string indicating the folder where docking data should be stored.}

\item{bio_assay_paths}{Vector indicating the absolute paths to the pdb files for the protein structures.}

\item{exhaust}{Exhaustiveness parameter for vina.}

\item{verbose}{Logical value. Print information or not.}

\item{path_save_control_matrix_path}{Path where the control matrix will be sotored or has been saved previouly.}
}
\value{

}
\description{
Performs docking completing a rectangle.
}
\examples{
\dontrun{
compute_by_square(bio_assay_pdbqt,drug_pdbqt_paths,output_path,bio_assay_paths,exhaust = 40,path_save_control_matrix_path,verbose = TRUE)
}

}
