% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{compute_exponential_sum}
\alias{compute_exponential_sum}
\title{compute_exponential_sum}
\usage{
compute_exponential_sum(quotient_dim)
}
\arguments{
\item{quotient_dim}{values of exponentiation function.}
}
\value{
returns the summation of the exponential.
}
\description{
Function that computes the summation of the exponential.
}
\examples{
\dontrun{
compute_exponential_sum(quotient_dim,k)
}
}
