% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_homology}
\alias{compute_homology}
\title{Compute horology for one point cloud.}
\usage{
compute_homology(alpha_coord, dimension_betti = 2)
}
\arguments{
\item{alpha_coord}{a matrix containing the alpha carbon coordinates.}

\item{dimension_betti}{the maximum bettin dimension to compute.}
}
\value{
returns the barcode of the input point cloud data.
}
\description{
Compute horology for one point cloud.
}
\examples{
\dontrun{
compute_homology(alpha_coord,2)
}
}
