% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_homology_multi}
\alias{compute_homology_multi}
\title{Computes homology for multiple alfa carbon coordinate matrices stored into a list.}
\usage{
compute_homology_multi(alpha_multi, dimension_betti = 2)
}
\arguments{
\item{alpha_multi}{}
}
\value{
returns a lists with the barcodes of the multiple point cloud data
}
\description{
Computes homology for multiple alfa carbon coordinate matrices stored into a list.
}
\examples{
\dontrun{
compute_homology_multi(alpha_multi,dimension_betti)
}
}
