% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{compute_k}
\alias{compute_k}
\title{compute_k}
\usage{
compute_k(quotient_ind, k)
}
\arguments{
\item{quotient_ind}{Value of the quotient in the exponentianted part of the function.}

\item{k}{Exponenet.}
}
\value{
Returns the exponentiated quotient.
}
\description{
This function exponentiates to k the quotient of the exponential.
}
\examples{
\dontrun{
compute_k(quotient_ind,k)
}
}
