% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_max_d}
\alias{compute_max_d}
\title{Computes max distance in a point cloud.}
\usage{
compute_max_d(alpha_coord, dimension_betti = 2)
}
\arguments{
\item{alpha_coord}{Matrix including the alpha carbon coordinates of a particular structure.}

\item{dimension_betti}{Maximum Betti dimension to compute.}
}
\value{
returns the diameter or maximum distance of the point cloud.
}
\description{
Computes max distance in a point cloud.
}
\examples{
\dontrun{
compute_max_d(alpha_coord)
}
}
