% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{compute_max_d_multi}
\alias{compute_max_d_multi}
\title{Compute maximum distance multi}
\usage{
compute_max_d_multi(alpha_multi)
}
\arguments{
\item{alpha_multi}{A list including a single or multiple matrices of alpha carbon coordinates generated using the extract_multiple_alpha function.}
}
\value{
Returns a list of max distances.
}
\description{
Compute the maximum distance for a list of alpha carbon coordinate matrices.
}
\examples{
\dontrun{
compute_max_d_multi(alpha_multi)
}
}
