% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{compute_pbf_at_x}
\alias{compute_pbf_at_x}
\title{compute_pbf_at_x}
\usage{
compute_pbf_at_x(x, barcode, k = 2)
}
\arguments{
\item{x}{x value}

\item{barcode}{a particular barcode.}

\item{k}{The value of exponent k}
}
\value{
Returns the pbf function at point x
}
\description{
compute_pbf_at_x
}
\examples{
\dontrun{
compute_pbf_at_x(x,barcode,k)
}
}
