% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{compute_pbf_for_curve}
\alias{compute_pbf_for_curve}
\title{compute_pbf_for_curve}
\usage{
compute_pbf_for_curve(vec_vals, barcode, k = 2, dim_betti = c(0, 1, 2))
}
\arguments{
\item{vec_vals}{vector with the x values to compute the PBF function.}

\item{barcode}{matrix with the barcode information generated by compute_homology or compute_homology_multi functions.}

\item{k}{Parameter k for exponentiation.}

\item{dim_betti}{Betti dimension to test.}
}
\value{
Returns PBF for a specified Betti dimension.
}
\description{
compute_pbf_for_curve
}
\examples{
\dontrun{
compute_pbf_for_curve(vec_vals,barcode,2,0)
}
}
