% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{compute_sim_barcode_pair}
\alias{compute_sim_barcode_pair}
\title{compute_sim_barcode_pair}
\usage{
compute_sim_barcode_pair(barcode_a, barcode_b, resolution = 0.01)
}
\arguments{
\item{barcode_a}{barcode structure a}

\item{barcode_b}{barcode structure b}

\item{resolution}{resolution parameter that indicates de width of the intervals used to approximate the function.}
}
\value{
returns a list with the similarities between the barcodes of
}
\description{
Compute the barcode betti similarity function for all the studied dimension in two given structures.
}
\examples{
\dontrun{
compute_sim_barcode_pair(barcode_a,barcode_b,resolution)
}
}
