% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{compute_sim_barcode_pair_aux}
\alias{compute_sim_barcode_pair_aux}
\title{compute_sim_barcode_pair_aux}
\usage{
compute_sim_barcode_pair_aux(pbf_a_one_dim, pbf_b_one_dim, range_x_axis)
}
\arguments{
\item{pbf_a_one_dim}{persistent betty function for a particular dimenson derived from barcode A}

\item{pbf_b_one_dim}{persistent betty function for a particular dimenson derived from barcode B}

\item{range_x_axis}{Comparison range.}
}
\value{
returns the persistent betty similarity for the dimensions of
}
\description{
compute_sim_barcode_pair_aux
}
\examples{
\dontrun{
compute_sim_barcode_pair_aux(pbf_a_one_dim,pbf_b_one_dim,range_x_axis)
}
}
