% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{compute_sim_one_dir_bar}
\alias{compute_sim_one_dir_bar}
\title{compute_sim_one_dir_bar}
\usage{
compute_sim_one_dir_bar(
  path_to_bcs,
  dir_out,
  n_cores = 30,
  tail_to_remove = ".Rda"
)
}
\arguments{
\item{path_to_bcs}{vector including the complete paths to all barcode files.}

\item{dir_out}{directory where the output data must be written.}

\item{n_cores}{Number of processores used in for each}

\item{tail_to_remove}{Tail tagg to remove from barcode name.}
}
\value{

}
\description{
compute_sim_one_dir_bar
}
\examples{
\dontrun{
compute_sim_one_dir_bar(path_to_bcs,dir_out,n_cores,tail_to_remove)
}
}
