% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{compute_sim_two_dir_bar}
\alias{compute_sim_two_dir_bar}
\title{compute_sim_two_dir_bar}
\usage{
compute_sim_two_dir_bar(dir_a, dir_b, dir_out, n_cores = 30)
}
\arguments{
\item{dir_a}{Paths to the first set of barcodes.}

\item{dir_b}{Paths to the second set of barcodes.}

\item{dir_out}{Output directory.}

\item{n_cores}{Number of processors for parallel computing.}
}
\value{

}
\description{
compute_sim_two_dir_bar
}
\examples{
\dontrun{
compute_sim_two_dir_bar(dir_a,dir_b,dir_out,n_cores)
}
}
