% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulate_PDB.R
\name{correct_pdbqt}
\alias{correct_pdbqt}
\title{correct_pdbqt}
\usage{
correct_pdbqt(x, out_file)
}
\arguments{
\item{x}{pdbqt file to correct.}

\item{out_file}{out file to save.}
}
\value{

}
\description{
correct_pdbqt
}
\examples{
\dontrun{
correct_pdbqt(x,out_file)
}
}
