% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docking_Functions.R
\name{create_config_pdb}
\alias{create_config_pdb}
\title{create_config_pdb}
\usage{
create_config_pdb(x)
}
\arguments{
\item{x}{pdb file.}
}
\value{

}
\description{
Create pdb configuration file.
}
\examples{
\dontrun{
create_config_pdb(x)
}

}
