% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docking_Functions.R
\name{create_vina_command_2}
\alias{create_vina_command_2}
\title{create_vina_command_2}
\usage{
create_vina_command_2(x, y, z, exhaust = 8)
}
\arguments{
\item{x}{Path to protein pdbqt.}

\item{y}{Path to ligand pdbqt.}

\item{z}{Output pathway.}

\item{exhaust}{Exhaustiveness parameter for vina.}

\item{w}{Path to receptor pdb files.}
}
\value{

}
\description{
Function to create vina commands.
}
\examples{
\dontrun{
create_vina_command(x,y,z,w,exhaust)
}

}
