% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{den_sub}
\alias{den_sub}
\title{den_sub}
\usage{
den_sub(barcode_dim)
}
\arguments{
\item{barcode_dim}{The barcode matrix for a particular dimension.}
}
\value{

}
\description{
Computes the substraction presennnn the exponentiated function of the PBF.
}
\examples{
\dontrun{
den_sub(barcode_dim)
}
}
