% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{download_pdb_structures}
\alias{download_pdb_structures}
\title{Download a large number of PDB files.}
\usage{
download_pdb_structures(pdb_ids_vec, path_to_PDB_folder, verbose = FALSE)
}
\arguments{
\item{pdb_ids_vec}{PDB ids for the structures to download.}

\item{path_to_PDB_folder}{Path to the folder were the structures must be downloaded.}
}
\value{

}
\description{
Download a large number of PDB files.
}
\examples{
\dontrun{
download_pdb_structures(pdb_ids_vec,path_to_PDB_folder)
}

}
