% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{element_wise_division}
\alias{element_wise_division}
\title{element_wise_division}
\usage{
element_wise_division(x, y)
}
\arguments{
\item{x}{a numeric vector x}

\item{y}{a numeric vertor y}
}
\value{
Returns a vector with the computed element-wise division.
}
\description{
Divides two vectors on an element-by-element basis.
}
\examples{
\dontrun{
element_wise_division(x,y)
}
}
