% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Barcodes.R
\name{extract_alpha_chains}
\alias{extract_alpha_chains}
\title{extract_alpha_chains}
\usage{
extract_alpha_chains(structure)
}
\arguments{
\item{structure}{}
}
\value{

}
\description{
Extract alpha chains from a pdb structure. Auxiliar function for compute_bc_round.
}
\examples{
\dontrun{
extract_alpha_chains(structure)
}
}
