% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{extract_and_clean_multi}
\alias{extract_and_clean_multi}
\title{Extract Bio And clean}
\usage{
extract_and_clean_multi(pdb_structures_paths, directory_to_write)
}
\arguments{
\item{pdb_structures_paths}{a vector containing the complete paths to the pdb structures.}

\item{directory_to_write}{a string including the destination folder where clean biological assemblies must be stored.}
}
\value{

}
\description{
Extracts biological assemblies from pdb structures and removes waters, ligans, and non-proteic atoms.
}
\examples{
\dontrun{
extract_and_clean_multi(pdb_structures_paths,directory_to_write)
}

}
