% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{extract_DrugBank_data}
\alias{extract_DrugBank_data}
\title{Extract Information from DrugBank}
\usage{
extract_DrugBank_data(
  path_to_drugbank_xml,
  path_to_data,
  path_to_uniprot_data,
  verbose = FALSE
)
}
\arguments{
\item{path_to_drugbank_xml}{Path to the file containing the xml format drugbank database.}

\item{path_to_data}{Path to the folder were the output data should be stored.}

\item{path_to_uniprot_data}{Path to the file containing the mapping between uniprot IDs and the PDB structures.}
}
\value{

}
\description{
Extract Information from DrugBank
}
\examples{
\dontrun{
extract_DrugBank_data(alpha_coord)
}

}
