% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulate_PDB.R
\name{extract_multiple_alpha}
\alias{extract_multiple_alpha}
\title{extract_multiple_alpha}
\usage{
extract_multiple_alpha(structures)
}
\arguments{
\item{structures}{A list of pdb structures loaded employing the bio3d read.pdb function.}
}
\value{
Returns a list with multiple matrices of alfa carbon coordinates. One for each structure in the input lits.
}
\description{
This function extracts the alpha carbon coordintes from a list containing multiple structures.
}
\examples{
\dontrun{
aux_for_extract_multiple_alpha(struc)
}
}
