% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Result_Mat.R
\name{format_matrices}
\alias{format_matrices}
\title{format_matrices}
\usage{
format_matrices(dist_mat, db_data)
}
\arguments{
\item{dist_mat}{list of similarity matrices computed by compute_sim_two_dir_bar function.}

\item{db_data}{Data frame containing the infromation from the reference database.}
}
\value{

}
\description{
format_matrices
}
\examples{
\dontrun{
format_matrices(dist_mat,db_data)
}
}
