% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{get_bio_assembly_information}
\alias{get_bio_assembly_information}
\title{get_bio_assembly_information}
\usage{
get_bio_assembly_information(path_to_bio_ass)
}
\arguments{
\item{path_to_bio_ass}{Pathway to biological assemblies folder.}
}
\value{

}
\description{
get_bio_assembly_information
}
\examples{
\dontrun{
get_bio_assembly_information(path_to_bio_ass)
}

}
