% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{get_bio_clean}
\alias{get_bio_clean}
\title{Get and clean biological assemblies.}
\usage{
get_bio_clean(pdb_object, directory_to_write, structure_name)
}
\arguments{
\item{pdb_object}{A pdb class object.}

\item{directory_to_write}{Path to the directory where clean biological units should be written.}

\item{structure_name}{Name of the original pdb strcuture.}
}
\value{

}
\description{
Get and clean biological assemblies.
}
\examples{
\dontrun{
get_bio_clean(pdb_object,directory_to_write,structure_name)
}

}
