% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_And_Search.R
\name{get_level_one}
\alias{get_level_one}
\title{get_level_one}
\usage{
get_level_one(dist_mat, n_clust)
}
\arguments{
\item{dist_mat}{Matrix of topological distances among biological assemblies.}

\item{n_clust}{Number of clusters to retrieve.}
}
\value{

}
\description{
Carries out PAM clustering of  a matrix of topological-based distances.
}
\examples{
\dontrun{
get_level_one(dist_mat,n_clust)
}
}
