% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_And_Search.R
\name{get_level_two}
\alias{get_level_two}
\title{get_level_two}
\usage{
get_level_two(df_data, dist_mat, n_clust)
}
\arguments{
\item{df_data}{Data.frame output from get_level_one}

\item{dist_mat}{Topological distance matrix}

\item{n_clust}{Number of clusters to produce in level two clustering.}
}
\value{

}
\description{
Carries out the second level of PAM clustering.
}
\examples{
\dontrun{
get_level_two(dist_mat,n_clust)
}
}
