% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docking_Functions.R
\name{get_paths_to_dock}
\alias{get_paths_to_dock}
\title{get_paths_to_dock}
\usage{
get_paths_to_dock(
  protein_test,
  drug_test,
  prot_estructures_prep,
  prot_estructures,
  drug_structures_prep,
  out_file
)
}
\arguments{
\item{protein_test}{biological assembly to dock}

\item{drug_test}{Drug to dock}

\item{prot_estructures_prep}{All prepared biological assemblies .pdbqt}

\item{prot_estructures}{All prepared biological assemblies .pdb}

\item{drug_structures_prep}{All prepared drug structures .pdbqt}

\item{out_file}{Output file}
}
\value{

}
\description{
get_paths_to_dock
}
\examples{
\dontrun{
get_paths_to_dock(protein_test,drug_test,prot_estructures_prep,prot_estructures,drug_structures_prep,out_file)
}

}
