% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{mount_matrix}
\alias{mount_matrix}
\title{mount_matrix}
\usage{
mount_matrix(dir_files, structures_ini, tail_to_remove = ".Rda")
}
\arguments{
\item{dir_files}{directory where the output intermediate files from compute_sim_one_dir_bar or compute_sim_two_dir_bar are placed}

\item{structures_ini}{original list of barcodes to compute used to compute similarities.}

\item{tail_to_remove}{character string to remove from the tail of the barcode file paths.}
}
\value{

}
\description{
Mounts distance matrix from partially computed distances.
}
\examples{
\dontrun{
mount_matrix(dir_files,structures_ini,tail_to_remove)
}
}
