% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docking_Functions.R
\name{parallel_drug_prep}
\alias{parallel_drug_prep}
\title{parallel_drug_prep}
\usage{
parallel_drug_prep(
  x,
 
    path_to_mk_prepare = "/home/antoniojr/anaconda3/envs/py36/scripts/mk_prepare_ligand.py",
  out_dir
)
}
\arguments{
\item{x}{Path to sdf drug file.}

\item{path_to_mk_prepare}{Absolute path to mk_prepare_ligand.py scrii}
}
\value{

}
\description{
Function to prepare drug structures for docking.
}
\examples{
\dontrun{
parallel_drug_prep(x,path_to_mk_prepare)
}

}
