% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsing_Vina_Logs.R
\name{parse_max_energy_vina}
\alias{parse_max_energy_vina}
\title{parse_max_energy_vina}
\usage{
parse_max_energy_vina(log_file_path)
}
\arguments{
\item{log_file_path}{}
}
\value{

}
\description{
Retrieves max binding energy from a vina log output file.
}
\examples{
\dontrun{
parse_max_energy_vina(log_file_path)
}
}
