% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{plot_barcodes}
\alias{plot_barcodes}
\title{plot_barcode}
\usage{
plot_barcodes(barcode_individual)
}
\arguments{
\item{barcode_individual}{A barcode matrix derived from the compute_homology or compute_homology_multi function.}
}
\value{
a plot object
}
\description{
Plotting barcode using the TDAstats function.
}
\examples{
\dontrun{
plot_dimensions(barcode_individual)
}
}
