% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docking_Functions.R
\name{prepare_protein_structure}
\alias{prepare_protein_structure}
\title{prepare_protein_structure}
\usage{
prepare_protein_structure(
  x,
 
    path_to_prepare_receptor = "/home/antoniojr/anaconda3/envs/py36/MGLToolsPckgs/AutoDockTools/Utilities24/prepare_receptor4.py",
  out_dir
)
}
\arguments{
\item{x}{paht to docking structure.}

\item{path_to_prepare_receptor}{absolute path to prepare_receptor4.py script}

\item{out_dir}{path to the directory where the output must be saved.}
}
\value{

}
\description{
prepares protein structures for docking.
}
\examples{
\dontrun{
prepare_protein_structure(x,path_to_prepare_receptor,out_dir)
}

}
