% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsing_Vina_Logs.R
\name{read_doc_res_from_list_dirs}
\alias{read_doc_res_from_list_dirs}
\title{read_doc_res_from_list_dirs}
\usage{
read_doc_res_from_list_dirs(x)
}
\arguments{
\item{x}{vector of directories to vina log files.}
}
\value{

}
\description{
Parses a list of docking results from a vector of directories.
}
\examples{
\dontrun{
read_doc_res_from_list_dirs(x)
}
}
