% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulate_PDB.R
\name{read_multi_pdb}
\alias{read_multi_pdb}
\title{Read multiple pdb files}
\usage{
read_multi_pdb(paths_to_structures)
}
\arguments{
\item{paths_to_structures}{directory were the structures are stored.}
}
\value{
a list of the loaded structures.
}
\description{
This function reads multiple pdb files containing multiple structures.
}
\examples{
\dontrun{
read_multi_pdb(paths_to_structures)
}
}
