% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{remove_ceros}
\alias{remove_ceros}
\title{remove_ceros}
\usage{
remove_ceros(x)
}
\arguments{
\item{x}{List of similarity matrices derived from the complete_matrices functions.}
}
\value{

}
\description{
Function that removes rows and columns that only contain zeros from the similiarity matrices.
}
\examples{
\dontrun{
remove_ceros(x)
}
}
