% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsing_Vina_Logs.R
\name{same_drug_diff_struc}
\alias{same_drug_diff_struc}
\title{same_drug_diff_struc}
\usage{
same_drug_diff_struc(x)
}
\arguments{
\item{x}{Complete dataset of docking results.}
}
\value{

}
\description{
same_drug_diff_struc
}
\examples{
\dontrun{
same_drug_diff_struc(x)
}
}
