% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{sample_structures_a}
\alias{sample_structures_a}
\title{sample_structures_a}
\usage{
sample_structures_a(x)
}
\arguments{
\item{x}{output dataframe from transform_sim_mat.}
}
\value{

}
\description{
Add group labels based on av similarity values.
}
\examples{
\dontrun{
sample_structures_a(x)
}
}
