% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{search_bcs}
\alias{search_bcs}
\title{search_bcs}
\usage{
search_bcs(
  bc_to_test,
  bc_to_test_name,
  level_one,
  level_two,
  dir_bcs = "/media/data/jaume/DocTDA/Data/Round_1/Barcodes",
  n_cores,
  lev_one_thr = 0.8,
  lev_two_thr = 0.85,
  lev_three_thr = 0.9
)
}
\arguments{
\item{bc_to_test}{Problem barcode.}

\item{bc_to_test_name}{Probelm barcode name.}

\item{level_one}{Level one clustering data of the database derived from get_level_one function}

\item{level_two}{Level two clustering data from the database derived from the get_level_two funciton.}

\item{dir_bcs}{Directory where the barcodes for the datbase structures are stored.}

\item{n_cores}{Number of cores available for parallel colmuting.}

\item{lev_one_thr}{Similarity threshold for level one clustering medioids.}

\item{lev_two_thr}{Similarity threshold for level two clustering medioids.}

\item{lev_three_thr}{Similarity thresuld for level three results.}
}
\value{

}
\description{
Given a specific barcode, the function carries out searches in the database and returns the barcodes displaying the higher simmilarity with the query barcode.
}
\examples{
\dontrun{
search_bcs(bc_to_test,bc_to_test_name,level_one,level_two,dir_bcs = "/media/data/jaume/DocTDA/Data/Round_1/Barcodes",n_cores,lev_one_thr = 0.8,lev_two_thr = 0.85,lev_three_thr = 0.90)
}
}
