% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{sum_num}
\alias{sum_num}
\title{sum_num}
\usage{
sum_num(barcode_dim)
}
\arguments{
\item{barcode_dim}{The barcode matrix for a particular dimension.}
}
\value{
Returns the fraction in the numerator of the PDF.
}
\description{
Computes the fraction of the numeratior of the exponentiated function in the PBF
}
\examples{
\dontrun{
sum_num(barcode_dim)
}
}
