% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_Functions.R
\name{transform_sim_mat}
\alias{transform_sim_mat}
\title{transform_sim_mat}
\usage{
transform_sim_mat(sim_matrices_comp)
}
\arguments{
\item{sim_matrices_comp}{list of similarity matrices}
}
\value{

}
\description{
Gets completed similarity matrices and generates a clean data.frame
}
\examples{
\dontrun{
transform_sim_mat(sim_matrices_comp)
}
}
