% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_PBF.R
\name{x_substraction}
\alias{x_substraction}
\title{x_sustraction}
\usage{
x_substraction(x, frac_num)
}
\arguments{
\item{x}{the value of x}

\item{frac_num}{the fraction numerator.}
}
\value{
Substraction value.
}
\description{
x_sustraction
}
\examples{
\dontrun{
den_sub(barcode_dim)
}
}
