% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_DB.R
\name{rename_drug_sdf}
\alias{rename_drug_sdf}
\title{rename_drug_sdf}
\usage{
rename_drug_sdf(path_to_sdfs)
}
\arguments{
\item{path_to_sdfs}{Path to the location of sdf 3D drug structures.}
}
\value{

}
\description{
Renames and SDF file generated using the split_sdf.py script using its DrugBank ID.
}
